fix bug with inlining scripts when webpack asset hash contains public path prefix
Some versions of Webpack have a compilation asset object of {'app.chunkHash.js': VALUE},
some have {'publicPath/app.chunkHash.js': VALUE}. Prior to this commit, the plugin
looked at a script tag's src value, stripped all prefixes from the asset's name, and
used that prefix-free name to directly reference the object in the compilation asset
object. This worked for prefix-free hashes but did not work when the asset object
contained the public path prefix.

This patch instead searches the asset object's keys for one which ends with the asset
name and uses that key to reference the asset; this accomodates both forms.